+++ src/features/CardDeck/CardSelectModal/index.tsx
@@ -15,7 +15,7 @@ import {
 import Image from 'next/image';
 import { useEffect, useState } from 'react';
 
-import Card from '../Card';
+import Card from '@/components/Card';
 import type { CardType } from '@/types/cards';
 
 type Props = { cards?: CardType[]; isOpen: boolean; onClose: () => void };
